backend/requirements.txt
Flask
Flask-Cors
yfinance
python-dotenv
pandas
numpy